# Allow library user to specify custom min and max-sizes