Added NewWithBoundaries(...) allowing the library user to specify custom min and max-size values. Used in Knoxite.