@fd0 2016-12-30
Cool, thanks! I've done a few stats, looks good:
```
$ benchstat before-go-1.7.4 after-go-1.7.4 
name                 old time/op   new time/op   delta
ChunkerWithSHA256-4    187ms ± 1%    188ms ± 1%   ~     (p=0.075 n=10+10)
Chunker-4             95.4ms ± 1%   95.9ms ± 2%   ~      (p=0.101 n=8+10)
NewChunker-4          74.6µs ± 8%   76.5µs ± 5%   ~     (p=0.315 n=10+10)

name                 old speed     new speed     delta
ChunkerWithSHA256-4  180MB/s ± 1%  179MB/s ± 1%   ~     (p=0.072 n=10+10)
Chunker-4            352MB/s ± 1%  350MB/s ± 2%   ~      (p=0.101 n=8+10)


$ benchstat before-go-1.8 after-go-1.8    
name                 old time/op   new time/op   delta
ChunkerWithSHA256-4    171ms ± 2%    172ms ± 1%     ~     (p=0.393 n=10+10)
Chunker-4             79.3ms ± 2%   80.0ms ± 1%   +0.92%   (p=0.017 n=10+9)
NewChunker-4          56.5µs ±12%   75.8µs ±30%  +34.14%   (p=0.004 n=9+10)

name                 old speed     new speed     delta
ChunkerWithSHA256-4  196MB/s ± 2%  196MB/s ± 1%     ~     (p=0.382 n=10+10)
Chunker-4            423MB/s ± 2%  419MB/s ± 1%   -0.92%   (p=0.017 n=10+9)
```